+++ DGTCentaurMods/game/eboard.py
@@ -11,7 +11,7 @@
 import serial
 import time
 from os.path import exists
-from dgt_centaur_mods.board import boardfunctions
+from DGTCentaurMods.board import boardfunctions
 import threading
 
 # https://github.com/well69/picochess-1/blob/master/test/dgtbrd-ruud.h
