+++ DGTCentaurMods/scripts/install.sh
